move janitor codes into agent/
