+++ agent/janitor/stats/counter.go
